Test classes where changed to implement BigDecimal instead of double

+++ src/test/java/com/example/roombooking/services/implementations/BookingServiceImplTest.java
@@ -15,6 +15,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -54,7 +55,7 @@ class BookingServiceImplTest {
 
     //Room
     private Long roomId = 2L;
-    private double price = 100.0;
+    private BigDecimal price = BigDecimal.valueOf(100.0);
     private RoomType roomType = new RoomType();
     private List<Booking> bookings = new ArrayList<>();
